feat(replays): Fix errors not being treated as a breadcrumb
Errors captured as a breadcrumb have a category of "sentry.event" with level of "error". This is different from Issue Details, where we create a fake "exception" crumb using the error event itself (not from breadcrumbs). I believe the difference here is also the crumb error is captured from client side, and the error event is our processed error from the backend. In the replays, it makes sense to show the client error.
